Layout Editor

+++ printing.py
@@ -15,6 +15,8 @@ def printImages(image_arr):
     logo_path = None
     image_marginx =None
     image_marginy =None
+    num_of_images =  len(image_arr)
+    logo_pos = None
     with open('./config/printing/layout.json', 'r') as layout_file:
         layout_data = json.load(layout_file)
         print(layout_data)
@@ -27,15 +29,17 @@ def printImages(image_arr):
         logo_path = layout_data['logo_path']
         image_marginx = layout_data['image_marginx']
         image_marginy = layout_data['image_marginy']
+        logo_pos = layout_data['logo_position']
+        
 
-
+    
 
 
 
 
 
     image_arr = ['./photos/{0}'.format(element) for element in image_arr]
-
+    image_arr.insert(logo_pos, logo_path)
    
     PHYSICALWIDTH = 110
     PHYSICALHEIGHT = 111
@@ -62,10 +66,10 @@ def printImages(image_arr):
   
 
     for i in range(len(image_arr)):
-        bmp = Image.open (image_arr[i]).convert('RGBA')
+        bmp = Image.open (image_arr[i])
         print(bmp.size[0])
         print(bmp.size[1])
-        if bmp.size[0] > bmp.size[1]:
+        if bmp.size[0] > bmp.size[1] or (i == logo_pos and logo_rotate==True):
             bmp = bmp.rotate (90, expand=True)
 
     
@@ -74,15 +78,10 @@ def printImages(image_arr):
         
         dib.draw (hDC.GetHandleOutput (), ((image_height*i) + marginy,marginx*2, image_height*i + image_height, int(printer_size[1]/2)-(marginx*2)))
         dib.draw (hDC.GetHandleOutput (), ((image_height*i) + marginy,int(printer_size[1]/2)+(marginx), image_height*i + image_height, int(printer_size[1])-(marginx*2)))
-        os.remove(image_arr[i])
+        if logo_pos != i:
+            os.remove(image_arr[i])
 
-    bmp = Image.open (logo_path).convert('RGBA')
-    if logo_rotate == True:
-            bmp = bmp.rotate (90, expand=True)
-    dib = ImageWin.Dib (bmp)
-    dib.draw (hDC.GetHandleOutput (), ((image_height*len(image_arr)) + image_marginy,image_marginx*2, image_height*len(image_arr) + logo_height, int(printer_size[1]/2)-(image_marginx*2)))
-    dib.draw (hDC.GetHandleOutput (), ((image_height*len(image_arr)) + image_marginy,int(printer_size[1]/2)+(image_marginx), image_height*len(image_arr) + logo_height, int(printer_size[1])-(image_marginx*2)))
-
+
